This commit was manufactured by cvs2svn to create tag 'oci_wchar_refactor_branch_pre_merge_in_4_24_06'.

+++ TAO/CIAO/examples/BasicSP/EC/EC_exec.h
@@ -62,7 +62,7 @@ namespace MyImpl
     int done_;
 
     /// The timer id we are waiting.
-    long tid_;
+    int tid_;
 
     EC_exec_i *pulse_callback_;
